chore: clean up legacy directories and consolidate db module
- Removed empty legacy directories (actions, components, db, lib)
- Consolidated database client to shared/lib/db.ts with full features
- Updated imports to use new shared database module
- Build compiles successfully
