--- src/db/index.ts
@@ -1,13 +1,0 @@
-import { Pool } from 'pg';
-
-if (!process.env.DATABASE_URL) {
-    throw new Error('DATABASE_URL environment variable is not set');
-}
-
-const pool = new Pool({
-    connectionString: process.env.DATABASE_URL,
-});
-
-export const query = (text: string, params?: unknown[]) => pool.query(text, params);
-
-export default pool;
